backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pgvector==0.3.6

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# Config
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# LLM & RAG
langchain>=0.3.13
langchain-google-genai>=2.0.0
langchain-community>=0.3.13
langchain-core>=0.3.30
google-genai>=1.0.0


# Document processing
pdfplumber==0.11.4
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==11.0.0
python-docx==1.1.2

# Storage
minio==7.2.11

# Utilities
httpx==0.28.1
tenacity==9.0.0
structlog==24.4.0
